requirements.txt
python-dotenv
pydantic
google-genai
stability-sdk